requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiofiles==23.2.1
aiogram==3.20.0.post0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiomysql==0.2.0
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.14.1
altair==5.1.1
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anthropic==0.52.0
anyio==3.7.1
APScheduler==3.10.4
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.1.0
Babel==2.9.1
bcrypt==4.3.0
cachetools==5.3.1
ccxt==4.0.72
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.2
decorator==4.4.2
distro==1.9.0
dlib==19.24.2
dnspython==2.7.0
easydict==1.10
ecdsa==0.19.1
email_validator==2.2.0
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.115.8
fastapi-cli==0.0.7
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
ffprobe==0.5
filelock==3.12.3
flatbuffers==25.2.10
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==3.33.1
gradio_client==0.5.0
greenlet==3.0.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.0
huggingface-hub==0.17.0
humanfriendly==10.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
imutils==0.5.4
iniconfig==2.1.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==3.9.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
linkify-it-py==2.0.2
loguru==0.7.0
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
magic-filter==1.0.12
Mako==1.3.10
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
namex==0.0.9
networkx==3.1
numpy==2.0.2
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.0
onnxruntime-silicon==1.15.0
openai==1.82.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.15.0
orjson==3.9.7
packaging==23.1
pandas==2.1.0
passlib==1.7.4
Pillow==10.0.0
pluggy==1.6.0
prettytable==3.9.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.5
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pydantic==2.10.6
pydantic-extra-types==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
replicate==1.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.10.2
rsa==4.9
schedule==1.2.0
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.9.1
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.12
tensorboard==2.19.0
tensorboard-data-server==0.7.1
tensorflow==2.19.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0
termcolor==2.3.0
tf==1.0.0
threadpoolctl==3.2.0
tifffile==2023.8.30
tomli==2.2.1
toolz==0.12.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
